requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-slugify==8.0.1
